pyproject.toml
[project]
name = "informe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[job-queue]>=22.5",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
